Extract common classes and constants to common.py; ignore type errors with boto3 clients
